+++ packages/chrome-devtools/src/utils/chrome/fast-refresh.ts
@@ -1,4 +1,7 @@
-import type { FederationRuntimePlugin } from '@module-federation/runtime/types';
+import type {
+  FederationRuntimePlugin,
+  Shared,
+} from '@module-federation/runtime/types';
 import { loadScript } from '@module-federation/sdk';
 
 import { isObject, getUnpkgUrl } from '../index';
@@ -29,53 +32,54 @@ const fastRefreshPlugin = (): FederationRuntimePlugin => {
           orderResolve = resolve;
         });
         Object.keys(shareInfo).forEach(async (share) => {
-          let get: () => any;
-          if (share === 'react') {
-            get = () =>
-              loadScript(
-                getUnpkgUrl(share, shareInfo[share].version) as string,
-                {
-                  attrs: { defer: true, async: false },
-                },
-              ).then(() => {
-                orderResolve();
-              });
-          }
-          if (share === 'react-dom') {
-            get = () =>
-              orderPromise.then(() =>
-                loadScript(
-                  getUnpkgUrl(share, shareInfo[share].version) as string,
-                  {
-                    attrs: { defer: true, async: false },
-                  },
-                ),
-              );
-          }
-          // @ts-expect-error
-          if (enableFastRefresh && typeof get === 'function') {
+          // @ts-expect-error legacy runtime shareInfo[share] is shared , and latest i shard[]
+          const sharedArr: Shared[] = Array.isArray(shareInfo[share])
+            ? shareInfo[share]
+            : [shareInfo[share]];
+
+          sharedArr.forEach((shared) => {
+            let get: () => any;
             if (share === 'react') {
-              shareInfo[share].get = async () => {
-                if (!window.React) {
-                  await get();
-                  console.warn(
-                    '[Module Federation HMR]: You are using Module Federation Devtools to debug online host, it will cause your project load Dev mode React and ReactDOM. If not in this mode, please disable it in Module Federation Devtools',
-                  );
-                }
-                shareInfo[share].lib = () => window.React;
-                return () => window.React;
-              };
+              get = () =>
+                loadScript(getUnpkgUrl(share, shared.version) as string, {
+                  attrs: { defer: true, async: false },
+                }).then(() => {
+                  orderResolve();
+                });
             }
             if (share === 'react-dom') {
-              shareInfo[share].get = async () => {
-                if (!window.ReactDOM) {
-                  await get();
-                }
-                shareInfo[share].lib = () => window.ReactDOM;
-                return () => window.ReactDOM;
-              };
+              get = () =>
+                orderPromise.then(() =>
+                  loadScript(getUnpkgUrl(share, shared.version) as string, {
+                    attrs: { defer: true, async: false },
+                  }),
+                );
             }
-          }
+            // @ts-expect-error
+            if (enableFastRefresh && typeof get === 'function') {
+              if (share === 'react') {
+                shared.get = async () => {
+                  if (!window.React) {
+                    await get();
+                    console.warn(
+                      '[Module Federation HMR]: You are using Module Federation Devtools to debug online host, it will cause your project load Dev mode React and ReactDOM. If not in this mode, please disable it in Module Federation Devtools',
+                    );
+                  }
+                  shared.lib = () => window.React;
+                  return () => window.React;
+                };
+              }
+              if (share === 'react-dom') {
+                shared.get = async () => {
+                  if (!window.ReactDOM) {
+                    await get();
+                  }
+                  shared.lib = () => window.ReactDOM;
+                  return () => window.ReactDOM;
+                };
+              }
+            }
+          });
         });
 
         return {